+++ API/Controllers/ContentsController.cs
@@ -4,6 +4,7 @@ using SimpleCMS.Application.Contents.Commands.DeleteContent;
 using SimpleCMS.Application.Contents.Commands.UpsertContent;
 using SimpleCMS.Application.Contents.Queries.GetContentDetail;
 using SimpleCMS.Application.Contents.Queries.GetContentsList;
+using System;
 using System.Threading.Tasks;
 
 namespace SimpleCMS.API.Controllers
@@ -11,9 +12,21 @@ namespace SimpleCMS.API.Controllers
     public class ContentsController : BaseController
     {
         [HttpGet]
-        public async Task<IActionResult> GetAll()
+        public async Task<IActionResult> GetAll(
+            [FromQuery] string searchTerm, [FromQuery] int topicId, [FromQuery] int categoryId,
+            [FromQuery] DateTime? createdAfter, [FromQuery] DateTime? modifiedAfter,
+            [FromQuery] int pageIndex = 1, [FromQuery] int pageSize = 10)
         {
-            return Ok(await Mediator.Send(new GetContentsListQuery()));
+            return Ok(await Mediator.Send(new GetContentsListQuery()
+            {
+                SearchTerm = searchTerm,
+                PageIndex = pageIndex,
+                PageSize = pageSize,
+                TopicId = topicId,
+                CategoryId = categoryId,
+                CreatedAfter = createdAfter,
+                ModifiedAfter = modifiedAfter
+            }));
         }
 
         [HttpGet("{id}")]