+++ Application/Contents/Queries/GetContentsList/GetContentsListQueryHandler.cs
@@ -1,11 +1,15 @@
 ﻿using AutoMapper;
 using AutoMapper.QueryableExtensions;
+using LinqKit;
 using MediatR;
-using Microsoft.EntityFrameworkCore;
 using SimpleCMS.Application.Common.Interfaces;
+using SimpleCMS.Application.Common.Collections;
 using SimpleCMS.Application.Contents.Queries.Common;
+using SimpleCMS.Domain.Entities;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using System.Data.Entity;
 
 namespace SimpleCMS.Application.Contents.Queries.GetContentsList
 {
@@ -22,14 +26,21 @@ namespace SimpleCMS.Application.Contents.Queries.GetContentsList
 
         public async Task<ContentsListVM> Handle(GetContentsListQuery request, CancellationToken cancellationToken)
         {
-            var contents = await _context.Contents.
-                ProjectTo<ContentDetailVM>(_mapper.ConfigurationProvider).
-                ToListAsync(cancellationToken);
+            var predicate = PredicateBuilder.New<Content>(true);
+
+            if (request.TopicId != 0) predicate = predicate.And(c => c.TopicId == request.TopicId);
+            if (request.CategoryId != 0) predicate = predicate.And(c => c.Topic.CategoryId == request.CategoryId);
+            if (request.CreatedAfter != null) predicate = predicate.And(c => c.Created >= request.CreatedAfter.Value);
+            if (request.ModifiedAfter != null) predicate = predicate.And(c => c.LastModified >= request.ModifiedAfter.Value);
+
+            var contents = _context.Contents.Where(predicate).AsNoTracking().ProjectTo<ContentDetailVM>(_mapper.ConfigurationProvider);
+
+            var paginatedContents = await PaginatedList<ContentDetailVM>.CreateAsync(contents, request.PageIndex, request.PageSize);
 
             var viewModel = new ContentsListVM()
             {
-                Contents = contents,
-                Count = contents.Count
+                Contents = paginatedContents,
+                Count = paginatedContents.Count()
             };
 
             return viewModel;