+++ client/src/components/EditHierarchyModal/HierarchyItemRowList.vue
@@ -14,6 +14,7 @@
       v-bind:key="item.id"
       v-on:selected="selectItem(item.id)"
       v-on:titleUpdated="updateItem"
+      v-on:onClickDelete="deleteItem"
     />
 
     <hierarchy-item-row
@@ -26,10 +27,7 @@
       v-on:blurred="selectItem(0)"
     />
 
-    <button
-      class="add-button"
-      v-on:click="selectItem(-1)"
-    >
+    <button class="add-button" v-on:click="selectItem(-1)">
       <i class="las la-plus" />
     </button>
   </div>
@@ -46,26 +44,29 @@ export default Vue.extend({
     initialSelectedItemId: Number,
     showAllOption: Boolean,
   },
-  data: function () {
+  data: function() {
     return {
       selectedItemId: this.initialSelectedItemId,
       newItem: { id: -1, title: "" },
     };
   },
   methods: {
-    selectItem: function (itemId: number) {
+    selectItem: function(itemId: number) {
       this.selectedItemId = itemId;
       this.$emit("selected", this.selectedItemId);
     },
-    updateItem: function (itemId: number, newTitle: string) {
+    updateItem: function(itemId: number, newTitle: string) {
       this.$emit("itemUpdated", itemId, newTitle);
     },
+    deleteItem: function(itemId: number){
+      this.$emit("itemDeleted", itemId);
+    }
   },
   components: {
     HierarchyItemRow,
   },
   watch: {
-    items: function (newVal) {
+    items: function(newVal) {
       //When new items are passed, reset the current selection
       this.selectItem(this.initialSelectedItemId);
     },