+++ client/src/models.ts
@@ -16,6 +16,14 @@ export interface Topic {
 export interface Category {
   id: number;
   title: string;
+  topics: Topic[];
+}
+
+export interface SearchQuery {
+  searchTerm: string;
+  filters: SearchFilters;
+  pageIndex: number;
+  pageSize: number;
 }
 
 export interface SearchFilters {
@@ -23,4 +31,17 @@ export interface SearchFilters {
   categoryId: number;
   createdWithinDays: number;
   updatedWithinDays: number;
+}
+
+export interface APIQueryCollection {
+  getAll: Function;
+  getById: Function;
+  upsert: Function;
+  delete: Function;
+}
+
+export interface SimpleAPIClient {
+  content: APIQueryCollection;
+  topics: APIQueryCollection;
+  categories: APIQueryCollection;
 }